+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Unreleased
 ----------
+Fixed
+^^^^^
+- Fix normalization and conversion events getting incorrectly bundled when using ``StructuralMapDiv.bundle_metadata``
 
 1.1.0 - 2025-02-26
 ------------------